requirements.txt
viam-sdk
yolov5